+++ scp/__init__.py
@@ -3,3 +3,4 @@ from .scp import SCP
 
 def setup(bot):
     bot.add_cog(SCP(bot))
+
