+++ scp/info.json
@@ -1,17 +1,9 @@
 {
-    "author": [
-      "Bobloy",
-      "SnappyDragon"
-    ],
-    "min_bot_version": "3.3.0",
-    "description": "Look up SCP articles. Warning: Some of them may be too creepy or gruesome.",
-    "hidden": false,
-    "install_msg": "You are now connected to the SCP database. You may now proceed to access the data using `[p]load scp`, then `[p]help SCP`",
-    "requirements": [],
-    "short": "Look up SCP articles.",
-    "end_user_data_statement": "This cog does not store any End User Data",
-    "tags": [
-      "bobloy",
-      "gruesom"
-    ]
-  }
+  "author": ["Danstr#1984", "Northwood Bot Development Team"],
+  "name" : "SCP",
+  "short" : "SCP Cog but with added **zest**",
+  "installmsg" : "Enjoy! if this cog fails to load and mentions a lack of pyscp, please consult the readme in the SCP cogs folder on our github.",
+  "end_user_data_statement" : "We store nothing, Fuck storing things *na-na-na-naa*",
+  "requirements": ["git+https://github.com/NorthWood-Cogs/pyscp"]
+  
+}
