--- scpV2/__init__.py
@@ -1,5 +1,0 @@
-from .scp import SCP
-
-
-def setup(bot):
-    bot.add_cog(SCP(bot))
