--- scpV2/info.json
@@ -1,9 +1,0 @@
-{
-    "author": ["Danstr#1984", "Northwood Bot Development Team"],
-    "name" : "SCP",
-    "short" : "SCP Cog but with added **zest**",
-    "installmsg" : "Enjoy! if this cog fails to load and mentions a lack of pyscp, please consult the readme in the SCP cogs folder on our github.",
-    "end_user_data_statement" : "We store nothing, Fuck storing things *na-na-na-naa*",
-    "requirements": ["git+https://github.com/NorthWood-Cogs/pyscp"]
-    
-}
